% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.long}
\alias{ex.long}
\title{A simulated data in long format}
\format{
A list with objects:
\describe{
  \item{y}{outcome vector (length n*p=1000)}
  \item{x}{predictor vector (length n*p=1000)}
  \item{id}{vector of the subject ids (length n*p=1000)}
  \item{class}{vector of the known classes of the subjects (length n=100)}
}
}
\usage{
ex.long
}
\description{
A simulated dataset with n=100 longitudinal measurements of length p=10.
}
\keyword{datasets}
