% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex}
\alias{ex}
\title{Simulated data}
\format{
A dataframe with variables:
\describe{
  \item{y}{outcome vector}
  \item{x}{predictor vector}
  \item{id}{subject id}
}
}
\usage{
ex
}
\description{
The simulated dataset has n=100 longitudinal measurements of length p=10.
}
\keyword{datasets}
