% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSample}
\alias{getSample}
\title{Data generation (this function uses normal distribution)}
\usage{
getSample(
  n = 100,
  p = 5,
  a = 1,
  b = 0.3,
  c = 0.3,
  pi = c(0.65, 0.35, 0),
  v = 1
)
}
\arguments{
\item{n}{number of individuals}

\item{p}{number of measurement points}

\item{a}{intercept (for all classes)}

\item{b}{slope (for classes 2 and 3)}

\item{c}{quadratic growth term (for class 3)}

\item{pi}{prior probabilities}

\item{v}{variance of random error}
}
\value{
A list containing the following components:
\describe{
\item{1}{data in long format (id, y, time, time^2)}
\item{2}{data in wide format (mostly for plotting purposes)}
\item{3}{classes}
}
}
\description{
\code{getSample} is used to ....
}
\details{
Here are the details of the function...
}
