% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.stepmixl_orig}
\alias{plot.stepmixl_orig}
\title{plot.stepmixl_orig}
\usage{
\method{plot}{stepmixl_orig}(x, ...)
}
\arguments{
\item{x}{an object of class stepmixl.}
}
\description{
plot.stepmixl_orig is used to plot the results of
function stepmixl_orig.
}
\details{
Here are the details of the function...
}
