% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{purity}
\alias{purity}
\title{Cluster purity}
\usage{
purity(classes, clusters)
}
\arguments{
\item{classes}{known classes}

\item{clusters}{clusters}
}
\description{
\code{purity} is used to ....
}
\details{
Here are the details of the function...
}
