% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepmixl.R
\name{stepmixl}
\alias{stepmixl}
\title{Application of Box-Cox transformation with flexmix}
\usage{
stepmixl(y, x, id, K, seed = .Random.seed, notrans = TRUE)
}
\arguments{
\item{y}{outcome vector}

\item{x}{predictor(s) as a vector or a matrix.}

\item{id}{identifier, unique to each subject.}

\item{K}{number of mixture components, single value or a vector.}

\item{seed}{RNG seed for reproducibility.}

\item{notrans}{if TRUE, analysis is done without transformation for comparison (TRUE by default)}
}
\value{
A list containing the following components:
\describe{
\item{models}{maximum likelihood solutions for different numbers of components. The numbers of components are given in K.}
\item{nt}{maximum likelihood solutions for different numbers of components without transformation, if notrans=TRUE}
\item{K}{number of mixture components}
\item{conv}{convergence of the stepFlexmix function. Equals 1 if converged.}
\item{iter}{number of iterations until convergence}
\item{lambda}{value of lambda}
\item{logL}{log-likelihood}
\item{AIC}{Akaike information criterion}
\item{BIC}{Bayesian information criterion}
\item{ICL}{Integrated completed likelihoo}
}
}
\description{
\code{stepmixl} is used to ....
}
\details{
Here are the details of the function...
}
\examples{
\dontrun{
library(scaledbc)
summary(ex)
res <- stepmixl(ex$y,ex$x,ex$id,K=1:5,seed=1)
plot(res)
summary(res)
summary(res,digits=7)
}
}
\references{
...
}
\seealso{
...
}
