% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepmixl2.R
\name{stepmixl2}
\alias{stepmixl2}
\title{Brute-force application of scaled Box-Cox transformation in flexmix}
\usage{
stepmixl2(y, x, id, lambdas, K, classes, data)
}
\arguments{
\item{y}{response variable (in long format).}

\item{x}{predictor(s) as a vector or a matrix (in long format).}

\item{id}{identifier unique to each subject (in long format).}

\item{lambdas}{sequence of lambda calculation points.}

\item{K}{number of mixture components, single value or a vector.}

\item{classes}{known classes of the subjects for computing cluster purity (optional).}

\item{data}{an optional data frame containing the variables y, x and id.}
}
\value{
A list containing the following components:
\describe{
\item{K}{number of mixture components}
\item{lambda}{best value of lambda found}
\item{AIC}{Akaike information criterion (lower is better)}
\item{BIC}{Bayesian information criterion (lower is better)}
\item{ICL}{Integrated completed likelihood (lower is better)}
\item{purity}{cluster purity, higher is better (if classes are known)}
\item{pass}{proportion of clusters that pass Shapiro test for normality of residuals at alpha=0.05}
\item{conv}{1 if algorithm converged before reaching maximum number of iterations, 0 otherwise}
\item{iter}{number of iterations}
\item{logL}{log-likelihood of model}
\item{models}{flexmix model objects fitted with best lambda value for each K}
}
}
\description{
\code{stepmixl2} is used to ....
}
\details{
Here are the details of the function...
}
\examples{
\dontrun{
library(scaledbc)
summary(ex)
res <- stepmixl(y,x,id,lambdas=seq(-2,2,0.05),K=1:5,data=ex.long)
plot(res)
summary(res)
summary(res,digits=7)
}
}
\references{
...
}
\seealso{
...
}
