% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepmixl_opt.R
\name{stepmixl_opt}
\alias{stepmixl_opt}
\title{Optimized application of Box-Cox transformation in flexmix}
\usage{
stepmixl_opt(y, x, id, K, classes)
}
\arguments{
\item{y}{response variable (in long format)}

\item{x}{predictor(s) as a vector or a matrix (in long format)}

\item{id}{identifier unique to each subject (in long format)}

\item{K}{number of mixture components, single value or a vector}

\item{classes}{known classes of the subjects for computing cluster purity (optional)}
}
\value{
A list containing the following components:
\describe{
\item{K}{number of mixture components}
\item{lambda}{best value of lambda found}
\item{AIC}{Akaike information criterion (lower is better)}
\item{BIC}{Bayesian information criterion (lower is better)}
\item{ICL}{Integrated completed likelihood (lower is better)}
\item{purity}{cluster purity, higher is better (if classes are known)}
\item{pass}{proportion of clusters that pass Shapiro test for normality of residuals at alpha=0.05}
\item{conv}{1 if algorithm converged before reaching maximum number of iterations, 0 otherwise}
\item{iter}{number of iterations}
\item{logL}{log-likelihood of model}
}
}
\description{
\code{stepmixl_opt} is used to ....
}
\details{
The method first checks 11 values of lambda
in range [-5,5] with interval 1.
The lambda with highest likelihood is selected
as a mid-point for the next iteration.
On the second iteration the range is of length 2
and the interval between values is 0.2.
The final iteration is done similarly with
a range of length 0.4 and interval 0.04.
}
\examples{
\dontrun{
library(scaledbc)
summary(ex)
res <- stepmixl_opt(ex.long$y,ex.long$x,ex.long$id,K=1:5)
plot(res)
summary(res)
summary(res,digits=7)
}
}
\references{
...
}
\seealso{
...
}
