% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summary.stepmixl}
\alias{summary.stepmixl}
\title{summary.stepmixl}
\usage{
summary.stepmixl(object, ..., digits = 6)
}
\arguments{
\item{object}{an object of class stepmixl.}

\item{...}{further arguments passed to or from other methods.}

\item{digits}{minimal number of significant digits.}
}
\description{
summary.stepmixl is used to print the summary of
an object of class stepmixl.
}
\details{
Here are the details of the function...
}
