% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summary.stepmixl_orig}
\alias{summary.stepmixl_orig}
\title{summary.stepmixl_orig}
\usage{
\method{summary}{stepmixl_orig}(object, ..., digits = 6)
}
\arguments{
\item{object}{an object of class stepmixl_orig.}

\item{...}{further arguments passed to or from other methods.}

\item{digits}{minimal number of significant digits.}
}
\description{
summary.stepmixl_orig is used to print the summary of
stepmixl_orig.
}
\details{
Here are the details of the function...
}
