% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test.resid}
\alias{test.resid}
\title{Performs a Shapiro test of normality for each cluster in a model and
returns the proportion of clusters that "pass" the test at alpha = 0.05}
\usage{
test.resid(model, y)
}
\arguments{
\item{model}{model}

\item{y}{y}
}
\description{
\code{test.resid} is used to ....
}
\details{
Here are the details of the function...
}
