% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tran}
\alias{tran}
\title{Scaled power transformation}
\usage{
tran(y, lambda)
}
\arguments{
\item{y}{y}

\item{lambda}{lambda}
}
\description{
\code{tran} is used to ....
}
\details{
Here are the details of the function...
}
