% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.wide}
\alias{ex.wide}
\title{Simulated data in wide format}
\format{
A list with objects:
\describe{
  \item{y}{outcome matrix (dimension 100x10)}
  \item{x}{predictor matrix (dimension 100x10)}
  \item{id}{vector of the subject ids (length n=100)}
  \item{class}{vector of the known classes of the subjects (length n=100)}
}
}
\usage{
ex.wide
}
\description{
The simulated dataset  has n=100 longitudinal measurements of length p=10.
}
\keyword{datasets}
