% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.stepmixl}
\alias{plot.stepmixl}
\title{plot.stepmixl}
\usage{
\method{plot}{stepmixl}(x, ...)
}
\arguments{
\item{x}{an object of class stepmixl.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
plot.stepmixl is used to plot the results of
an object of class stepmixl.
}
\details{
Here are the details of the function...
}
