% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepmixl_opt.R
\name{stepmixl_opt}
\alias{stepmixl_opt}
\title{Optimized application of Box-Cox transformation in flexmix}
\usage{
stepmixl_opt(y, x, id, K, classes, data)
}
\arguments{
\item{y}{response variable (in long format)}

\item{x}{predictor(s) as a vector or a matrix (in long format)}

\item{id}{identifier unique to each subject (in long format)}

\item{K}{number of mixture components, single value or a vector}

\item{classes}{known classes of the subjects for computing cluster purity (optional)}

\item{data}{a data frame containing the variables y, x and id (optional)}
}
\value{
A list with objects:
\describe{
\item{K}{number of mixture components}
\item{lambda}{best value of lambda found}
\item{AIC}{Akaike information criterion (lower is better)}
\item{BIC}{Bayesian information criterion (lower is better)}
\item{ICL}{Integrated completed likelihood (lower is better)}
\item{purity}{cluster purity, higher is better (if classes are known)}
\item{pass}{proportion of clusters that pass Shapiro test for normality of residuals at alpha=0.05}
\item{conv}{1 if algorithm converged before reaching maximum number of iterations, 0 otherwise}
\item{iter}{number of iterations}
\item{logL}{log-likelihood of model}
}
}
\description{
\code{stepmixl_opt} is used to ....
}
\details{
The method first checks 11 values of lambda
in range [-5,5] with interval 1.
The lambda with highest likelihood is selected
as a mid-point for the next iteration.
On the second iteration the range is of length 2
and the interval between values is 0.2.
The final iteration is done similarly with
a range of length 0.4 and interval 0.04.
}
\examples{
\dontrun{
library(scaledbc)
res <- stepmixl_opt(y,x,id,K=1:5,data=ex.long)
plot(res)
summary(res)
summary(res,digits=7)
}
}
\references{
Grün, B.  and Leisch, F. (2023), \emph{flexmix: Flexible Mixture Modeling}. R package version 2.3-19,
\url{https://CRAN.R-project.org/package=flexmix}.

Leisch, F. (2004). FlexMix: A General Framework for Finite Mixture Models and Latent Class
Regression in R. \emph{Journal of Statistical Software}, \strong{11}(8), 1-18,
\url{https://doi.org/10.18637/jss.v011.i08}.

Grün, B. and Leisch, F. (2007). Fitting Finite Mixtures of Generalized Linear Regressions in R,
\emph{Computational Statistics & Data Analysis}, \strong{51}, 5247–5252,
\url{https://doi.org/10.1016/j.csda.2006.08.014}.

Grün, B. and Leisch, F. (2008). FlexMix Version 2: Finite Mixtures with Concomitant Variables
and Varying and Constant Parameters. \emph{Journal of Statistical Software}, \strong{28}(4), 1-35,
\url{https://doi.org/10.18637/jss.v028.i04}.
}
